comments, error handling

+++ operator/webserver.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 from flask import Flask, request, abort, render_template
-import subprocess, time, signal, sys, os, json
+import subprocess, time, signal, sys, os, errno, json
 from datetime import datetime
 from threading import Timer
 
@@ -28,53 +28,77 @@ with open('../config/team_config.json', 'r') as file:
 def index():
     return render_template('index.html', team_config=team_config, logs=logs)
 
-# curl --header "Content-Type: application/json" --request POST --data '{"master": "http://localhost:11311", "mission": "trackdrive"}' http://localhost:5000/mission/selectcar
 @app.route('/mission/start', methods=['POST'])
 def mission_start():
+    # Abort if empty request
     if request.json is None or request.json['id'] is None or request.json['mission'] is None:
         return abort(400)    
 
+    # Get team config
     teamId = request.json['id']
     mission = request.json['mission']
     for obj in team_config: 
         if obj['id'] == teamId: team = obj
 
+    # Set ROS MASTER
     procenv = os.environ.copy()
     procenv["ROS_MASTER_URI"] = team['master']
 
+    # Launch ROS bridge
     global interfaceprocess, log_file
     interfaceprocess = subprocess.Popen(['roslaunch', 'fsds_ros_bridge', 'fsds_ros_bridge.launch', 'mission:={}'.format(mission)], env=procenv)   
 
+    # Create log message
     log = '{}: {}'.format(str(datetime.now()), 'Mission started.')
     logs.append(log)
 
-    log_file = open('logs/{}_{}_{}.txt'.format(team['name'].lower().replace(' ', '-'), mission, str(datetime.now().strftime("%d-%m-%Y_%H:%M:%S"))), 'w')
+    # Create log file. Create logs directory if it does not exist
+    filename = 'logs/{}_{}_{}.txt'.format(team['name'].lower().replace(' ', '-'), mission, str(datetime.now().strftime("%d-%m-%Y_%H:%M:%S")))
+    if not os.path.exists(os.path.dirname(filename)):
+        try:
+            os.makedirs(os.path.dirname(filename))
+        except OSError as exc: # Guard against race condition
+            if exc.errno != errno.EEXIST:
+                raise
+    
+    # Write to log file
+    log_file = open(filename, 'w')
     log_file.write(log + '\n')
 
     return {'response': log}
 
 @app.route('/mission/stop', methods=['POST'])
 def mission_stop():
-    # check if previous process is still running
+    # Abort if ROS bridge is not running
+    if interfaceprocess is None:
+        return abort(400)
+
+    # Check if previous process is still running
     if interfaceprocess is not None and interfaceprocess.poll() is None:
-        # try to stop it gracefully. SIGINT is the equivilant to doing ctrl-c
+        # Try to stop it gracefully. SIGINT is the equivilant to doing ctrl-c
         interfaceprocess.send_signal(signal.SIGINT)
         time.sleep(3)
-        # still running?
+        # Still running?
         if interfaceprocess.poll() is None:
-            # kill it with fire
+            # Kill it with fire
             interfaceprocess.terminate()
-            # wait for it to finish
+            # Wait for it to finish
             interfaceprocess.wait()
 
+        global interfaceprocess
+        interfaceprocess = None
+
+    # Brake car
     car_controls.brake = 1
     client.setCarControls(car_controls)
     car_controls.brake = 0 #remove brake
 
+    # Create log message
     log = '{}: {}'.format(str(datetime.now()), 'Mission stopped.')
     logs.append(log)
-    del logs[:]
+    del logs[:] # Clear logs
 
+    # Write logs to file
     log_file.write(log + '\n')
     log_file.close()
 
@@ -82,8 +106,10 @@ def mission_stop():
 
 @app.route('/mission/reset', methods=['POST'])
 def mission_reset():
+    # Reset simulator
     client.reset()
 
+    # Create log message and write to file
     log = '{}: {}'.format(str(datetime.now()), 'Car reset.')
     logs.append(log)
     log_file.write(log + '\n')
